requirements_ignore.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
click @ file:///tmp/build/80754af9/click_1646056590078/work
dill==0.3.6
etils==0.9.0
Flask @ file:///tmp/abs_f02y1lyx6o/croots/recipe/flask_1658870763686/work
flatbuffers==22.10.26
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
promise==2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.1
tensorflow-datasets==4.7.0
tensorflow-directml-plugin==0.2.0.dev221020
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.11.0
termcolor==2.1.0
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
